refactor button components
