+++ src/UI/Buttons/ButtonBase/index.tsx
@@ -1,0 +1,10 @@
+import styled from 'styled-components'
+
+export const ButtonBase = styled.button`
+  box-sizing: border-box;
+  text-transform: none;
+  border-radius: 4px;
+  border: none;
+  box-shadow: 0px 2px 8px rgba(0, 0, 0, 0.135216);
+  cursor: pointer;
+`
