+++ src/UI/Buttons/ButtonConfirmed/ButtonConfirmedStyle.tsx
@@ -1,0 +1,14 @@
+import styled from 'styled-components'
+import { lighten } from 'polished'
+import { Base } from '../Base'
+
+export const ButtonConfirmedStyle = styled(Base)`
+  background-color: ${({ theme }) => lighten(0.5, theme.green1)};
+  color: ${({ theme }) => theme.green1};
+  border: 1px solid ${({ theme }) => theme.green1};
+
+  &:disabled {
+    opacity: 50%;
+    cursor: auto;
+  }
+`
