+++ src/UI/Buttons/ButtonGray/index.tsx
@@ -1,0 +1,19 @@
+import styled from 'styled-components'
+import { darken } from 'polished'
+import { Base } from '../Base'
+
+export const ButtonGray = styled(Base)`
+  background-color: ${({ theme }) => theme.bg3};
+  color: ${({ theme }) => theme.text2};
+  font-size: 16px;
+  font-weight: 500;
+  &:focus {
+    background-color: ${({ theme, disabled }) => !disabled && darken(0.05, theme.bg4)};
+  }
+  &:hover {
+    background-color: ${({ theme, disabled }) => !disabled && darken(0.05, theme.bg4)};
+  }
+  &:active {
+    background-color: ${({ theme, disabled }) => !disabled && darken(0.1, theme.bg4)};
+  }
+`
