+++ src/UI/Buttons/ButtonMintBack/index.tsx
@@ -1,0 +1,7 @@
+import styled from 'styled-components'
+import { ButtonDraft } from '../ButtonDraft'
+
+export const ButtonMintBack = styled(ButtonDraft)`
+  color: #3f3f3f;
+  margin-right: 0;
+`
