+++ src/UI/Buttons/ButtonPink/index.tsx
@@ -1,0 +1,25 @@
+import styled from 'styled-components'
+import { darken } from 'polished'
+import { Base } from '../Base'
+
+export const ButtonPink = styled(Base)`
+  background-color: ${({ theme }) => theme.primary1};
+  color: white;
+
+  &:focus {
+    box-shadow: 0 0 0 1pt ${({ theme }) => darken(0.05, theme.primary1)};
+    background-color: ${({ theme }) => darken(0.05, theme.primary1)};
+  }
+  &:hover {
+    background-color: ${({ theme }) => darken(0.05, theme.primary1)};
+  }
+  &:active {
+    box-shadow: 0 0 0 1pt ${({ theme }) => darken(0.1, theme.primary1)};
+    background-color: ${({ theme }) => darken(0.1, theme.primary1)};
+  }
+  &:disabled {
+    background-color: ${({ theme }) => theme.primary1};
+    opacity: 50%;
+    cursor: auto;
+  }
+`
