+++ src/UI/Buttons/ButtonSearch/index.tsx
@@ -1,0 +1,8 @@
+import styled from 'styled-components'
+import { ButtonPrimary } from '../ButtonPrimary'
+
+export const ButtonSearch = styled(ButtonPrimary)`
+  width: 8.5vw;
+  border-radius: 0px 4px 4px 0px;
+  font-size: 1.125rem;
+`
