+++ src/UI/Buttons/ButtonSecondary/index.tsx
@@ -1,0 +1,30 @@
+import styled from 'styled-components'
+import { Base } from '../Base'
+
+export const ButtonSecondary = styled(Base)`
+  border: 1px solid ${({ theme }) => theme.primary4};
+  color: ${({ theme }) => theme.primary1};
+  background-color: transparent;
+  font-size: 16px;
+  border-radius: 12px;
+  padding: ${({ padding }) => (padding ? padding : '10px')};
+
+  &:focus {
+    box-shadow: 0 0 0 1pt ${({ theme }) => theme.primary4};
+    border: 1px solid ${({ theme }) => theme.primary3};
+  }
+  &:hover {
+    border: 1px solid ${({ theme }) => theme.primary3};
+  }
+  &:active {
+    box-shadow: 0 0 0 1pt ${({ theme }) => theme.primary4};
+    border: 1px solid ${({ theme }) => theme.primary3};
+  }
+  &:disabled {
+    opacity: 50%;
+    cursor: auto;
+  }
+  a:hover {
+    text-decoration: none;
+  }
+`
