+++ src/UI/Buttons/ButtonUNIGradient/index.tsx
@@ -1,0 +1,22 @@
+import styled from 'styled-components'
+import { CustomButtonPrimary } from '../CustomButtonPrimary'
+
+export const ButtonUNIGradient = styled(CustomButtonPrimary)`
+  color: white;
+  padding: 4px 8px;
+  height: 36px;
+  font-weight: 500;
+  background-color: ${({ theme }) => theme.bg3};
+  background: radial-gradient(174.47% 188.91% at 1.84% 0%, #ff007a 0%, #2172e5 100%), #edeef2;
+  width: fit-content;
+  position: relative;
+  cursor: pointer;
+  border: none;
+  white-space: no-wrap;
+  :hover {
+    opacity: 0.8;
+  }
+  :active {
+    opacity: 0.9;
+  }
+`
