+++ src/UI/Buttons/DelistButton/index.tsx
@@ -1,0 +1,13 @@
+import styled from 'styled-components'
+
+export const DelistButton = styled.button`
+  width: 168px;
+  height: 50px;
+  border: none;
+  background: #fbfbfb;
+  border-radius: 4px;
+  color: #858585;
+  font-family: Roboto;
+  font-size: 1rem;
+  cursor: pointer;
+`
