+++ src/components/NFTcard/nftcard.tsx
@@ -2,7 +2,7 @@ import React from 'react'
 import { Media, CardImg, Card, Price, Text, Actions, Bid } from './nftcard.styles'
 import { useTranslation } from 'react-i18next'
 import ButtonWishlist from 'UI/Buttons/ButtonWishlist'
-import { NftButton } from 'UI/Buttons'
+import { NftButton } from 'UI/Buttons/NftButton'
 import { AuctionNFT } from 'services/models/AuctionNFT'
 import uriToHttp from 'utils/uriToHttp'
 import Amount from 'components/NFTSummary/Amount'
