+++ src/components/OffMarket/OffMarket.tsx
@@ -9,7 +9,9 @@ import uriToHttp from 'utils/uriToHttp'
 import Card from '../../UI/Card'
 import { ImageContainer, Divider, TextContainer, TagRow } from 'components/InMarketAsset/InMarket.styles'
 import { Footer, TopTitle } from './OffMarket.styles'
-import { ButtonDraft, ButtonMint, ButtonMintBack } from 'UI/Buttons'
+import { ButtonMintBack } from 'UI/Buttons/ButtonMintBack'
+import { ButtonDraft } from 'UI/Buttons/ButtonDraft'
+import { ButtonMint } from 'UI/Buttons/ButtonMint'
 import { useTranslation } from 'react-i18next'
 import Vector from '../../assets/images/Vector.png'
 import StringModifier from 'utils/StringSplice'
