+++ src/components/TransactionConfirmationModal/index.tsx
@@ -9,7 +9,7 @@ import { Text } from 'rebass'
 import { CloseIcon, CustomLightSpinner } from '../../theme/components'
 import { RowBetween } from 'UI/Row'
 import { AlertTriangle, ArrowUpCircle } from 'react-feather'
-import { CustomButtonPrimary as ButtonPrimary } from '../../UI/Buttons'
+import { CustomButtonPrimary as ButtonPrimary } from '../../UI/Buttons/CustomButtonPrimary'
 import { AutoColumn, ColumnCenter } from 'UI/Column'
 import Circle from '../../assets/images/blue-loader.svg'
 import { getEtherscanLink } from '../../utils'
