+++ src/components/WishList/WishCard.tsx
@@ -4,7 +4,7 @@ import uriToHttp from 'utils/uriToHttp'
 import { WishListCard, TextContainer, RemoveContainer, ImgDIV, Counter, CounterContainer } from './WishList.styles'
 import Text from '../../UI/Text'
 import { useRemoveWishlistItem } from 'state/user/hooks'
-import { RemoveWishList } from 'UI/Buttons'
+import { RemoveWishList } from 'UI/Buttons/RemoveWishList'
 import Remove from '../../assets/images/Remove.png'
 import { useCountDownTimer } from 'hooks/countDownTimer'
 import { useTranslation } from 'react-i18next'
